+++ modules/goals/types/Seller.ts
@@ -1,0 +1,4 @@
+export type Seller = {
+  co_ven: string;
+  des_ven: string;
+}